Cambio de estilos a inputs y refactor
