--- src/components/Description/Description.style.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const DescriptionContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-
-  flex-grow: 1;
-
-  margin: ${p => p.theme.sizes.tiny};
-`
