+++ src/components/Description/Description.tsx
@@ -1,5 +1,15 @@
 import * as React from 'react'
-import { DescriptionContainer } from './Description.style'
+import styled from 'styled-components'
+
+export const DescriptionContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+
+  flex-grow: 1;
+
+  margin: ${p => p.theme.sizes.tiny};
+`
 
 interface Props {
   DescriptionText: string
