--- src/components/Edit/Edit.style.tsx
@@ -1,48 +1,0 @@
-import styled from 'styled-components'
-
-export const EditButtonStyle = styled.button`
-  float: right;
-
-  opacity: 0;
-
-  height: ${p => p.theme.sizes.large};
-  width: ${p => p.theme.sizes.large};
-
-  position: absolute;
-
-  right: 0;
-
-  z-index: 5;
-
-  color: blue;
-
-  cursor: pointer;
-
-  margin: 3px;
-  padding: ${p => p.theme.sizes.xTiny};
-
-  border: 1px solid #f0f0f0;
-  border-radius: 50%;
-
-  background-color: #f0f0f0;
-
-  box-shadow: 0 ${p => p.theme.sizes.tiny} ${p => p.theme.sizes.small}
-      rgba(0, 0, 0, 0.19),
-    0 6px 6px rgba(0, 0, 0, 0.23);
-
-  outline: none;
-  :hover {
-    opacity: 0.8;
-  }
-
-  @media only screen and (max-width: 769px) {
-    opacity: 0.3;
-    background-color: transparent;
-    border: none;
-    box-shadow: none;
-  }
-`
-
-export const EditImg = styled.img`
-  background-color: transparent;
-`
