+++ src/views/Home/Home.styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { Button } from '$components/Button'
+
+export const HamburgerMenuButtom = styled(Button)`
+  margin: 0;
+`
