+++ src/views/Home/components/AsideHome/AsideHome.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import { AsideWrapper } from './AsideHome.style'
+import { AsideWrapper } from './AsideHome.styles'
 
-export const AsideHome: React.SFC = ({children}) =>
-  (<AsideWrapper>{children}</AsideWrapper>)
-
+export const AsideHome: React.SFC = ({ children }) => (
+  <AsideWrapper>{children}</AsideWrapper>
+)
