+++ src/views/Home/components/HeaderHome/HeaderHome.styles.ts
@@ -3,12 +3,10 @@ import styled from 'styled-components'
 export const HeaderContainer = styled.div`
   position: fixed;
   display: flex;
-  flex-wrap: wrap;
+  flex-direction: row;
 
   justify-content: center;
   align-items: center;
-
-  margin: 0 auto;
 
   height: ${p => p.theme.sizes.huge};
   width: 100%;
@@ -17,5 +15,4 @@ export const HeaderContainer = styled.div`
   box-shadow: 0 4px 2px -2px ${p => p.theme.colors.darkGray};
 
   z-index: 10;
-
 `
