+++ src/views/Home/components/HeaderHome/HeaderHome.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react'
-import { HeaderContainer } from './HeaderHome.style'
+import { HeaderContainer } from './HeaderHome.styles'
 
-export const HeaderHome: React.SFC = ({children}) => (<HeaderContainer>{children}</HeaderContainer>)
-
+export const HeaderHome: React.SFC = ({ children }) => (
+  <HeaderContainer>{children}</HeaderContainer>
+)
