+++ src/views/Home/components/MainHome/MainHome.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react'
-import { MainWrapper } from './MainHome.style'
+import { MainWrapper } from './MainHome.styles'
 
-export const MainHome: React.SFC = ({children}) => <MainWrapper>{children}</MainWrapper>
-
+export const MainHome: React.SFC = ({ children }) => (
+  <MainWrapper>{children}</MainWrapper>
+)
