+++ src/views/Home/components/SectionHome/SectionHome.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import { SectionWrapper } from './SectionHome.style'
+import { SectionWrapper } from './SectionHome.styles'
 
-export const SectionHome: React.SFC= ({children}) => <SectionWrapper>{children}</SectionWrapper>
-  
-
+export const SectionHome: React.SFC = ({ children }) => (
+  <SectionWrapper>{children}</SectionWrapper>
+)
