+++ src/views/Login/Login.view.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react'
 import { LogInForm } from './components'
-import { LoginContainer } from './Login.style'
+import { LoginContainer } from './Login.styles'
 
-export const Login: React.SFC = () => 
-(
-      <LoginContainer>
-        <LogInForm />
-      </LoginContainer>
+export const Login: React.SFC = () => (
+  <LoginContainer>
+    <LogInForm />
+  </LoginContainer>
 )
-  
-
